Create .coaches__col-bottom-margin class for consistent spacing of coach info...
Minor consistency adjustments for coaches section on mobile devices:

Add {margin-bottom: 20px} to each coach column. Coach thumbnail images are clickable links and needed a little spacing in mobile view.

The softball page, which contains 2 rows of coach info, had a margin between the two rows in mobile views.
This created an inconsistent look of 3 coaches with no bottom margins, followed by a single coach with a gap caused by a margin:
Removed the margin that was on the bottom row of coaches on the softball page.