Create a second SASS compile gulp-task to compile a second stylesheet for CMS
Made another gulp task called contentSass that compiles a small content.css file that the CloudCannon CMS uses
This makes it so our custom heading and paragraph styles are available to content editors in the styles dropdown.

This has only been implimented in the fitness-center page.
Waiting for further testing and apporval before implimenting globaly.
Go to Fitness Center page in CloudCannon visual editor to see how this works

Options for the visual editor are defined in this pages' front-matter